+++ CHANGELOG.md
@@ -1,9 +1,21 @@
 # Change Log
 
-All notable changes to the "wmedrano" extension will be documented in this file.
+All notable changes to the "wmedrano" extension will be documented in this
+file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on
+how to structure this file.
 
 ## [Unreleased]
 
-- Initial release
+### 2.0.0
+
+Introduce `Write Wmedrano Config` command.
+
+### 1.0.0
+
+Rename `extension.wmedrano.quickGrep` command to `extension.wmedrano.quickSearch`.
+
+### 0.0.1
+
+Initial release of wmedrano vscode extensions. Implements `Quick Search`.